requirements.txt
openai~=0.27.2
tiktoken~=0.3.2
discord-py-interactions~=5.0.1